pyproject.toml
[project]
name = "py-music163"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "openai>=2.29.0",
    "pycryptodome>=3.23.0",
    "pyncm>=1.8.1",
    "requests>=2.32.4",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
